move HuggingFaceLocalGenerator to right directory
