+++ haystack/preview/components/generators/__init__.py
@@ -1,4 +1,4 @@
-from haystack.preview.components.generators.hugging_face.hugging_face_local import HuggingFaceLocalGenerator
+from haystack.preview.components.generators.hugging_face_local import HuggingFaceLocalGenerator
 from haystack.preview.components.generators.hugging_face_tgi import HuggingFaceTGIGenerator
 from haystack.preview.components.generators.openai import GPTGenerator
