+++ test/preview/components/generators/test_hugging_face_local_generator.py
@@ -4,10 +4,7 @@ from unittest.mock import patch, Mock
 import pytest
 import torch
 
-from haystack.preview.components.generators.hugging_face.hugging_face_local import (
-    HuggingFaceLocalGenerator,
-    StopWordsCriteria,
-)
+from haystack.preview.components.generators.hugging_face_local import HuggingFaceLocalGenerator, StopWordsCriteria
 
 
 class TestHuggingFaceLocalGenerator:
